+++ src/main/java/com/example/demo/Service/PedidoService.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface PedidoService {
-    public Pedido getById(BigInteger id);
+    public Pedido getById(Long id);
     public List<Pedido> getAll();
     public Pedido save(Pedido pedido);
-    public Pedido update(BigInteger id, Pedido pedido);
-    public void delete(BigInteger id);
+    public Pedido update(Long id, Pedido pedido);
+    public void delete(Long id);
 }
